handshaker fixes:
- return result before filtering
- remove fork version filter
- extract handshake filters
